feat: Add product validation endpoint
This commit adds a new endpoint to the ProductsController for validating products. The endpoint is triggered by a message pattern with the command 'validate_products' and accepts an array of product IDs. The method in the ProductsService performs a lookup for the provided IDs and throws an error if any of the products are not found. If all products are found, the method returns the list of products.